chore: Move split to own transformer
Moves extractNestedObject to common
