+++ common.go
@@ -242,3 +242,18 @@ func Assert(x bool, v ...interface{}) {
 		panic(fmt.Sprintf("%s %s", out, fmt.Sprint(v...)))
 	}
 }
+
+// ExtractNestedObject extracts an object from a nested object structure. All intermediate objects has to map[string]interface{}
+func ExtractNestedObject(object map[string]interface{}, keys []string) (map[string]interface{}, error) {
+	if len(keys) == 1 {
+		return object, nil
+	}
+
+	next, ok := object[keys[0]].(map[string]interface{})
+
+	if !ok {
+		return nil, Error{Reason: "Failed to cast nested object to map[string]interface{}"}
+	}
+
+	return ExtractNestedObject(next, keys[1:])
+}
